+++ Release Notes.md
@@ -12,7 +12,8 @@ This will make the things in this library easier to understand and document, sin
 
 ### ✨ New features
 
-* `View+PersistentSystemOverlays` adds a view extension for only hiding persistent system overlays if the feature is available. 
+* `View+PersistentSystemOverlays` adds a view extension for only hiding persistent system overlays if the feature is available.
+* `View+Searchable` adds a conditional searchable view extension.